Ensure user is always created on the Stream client
The `findOneOrCreate` method always creates the user if one doesn't
exist which meant that the code for creating the user on the channel
never ran since it was in a conditional that only executed if the user
is not found in the database.

Here, the `fineOne` method is used first to check if the user exists. If
not the user is created on the database, and subsequently on the Stream
instance.